#137 Major reorganisation of these Tokens.
