+++ Util/src/pParse/NatBase32Token.scala
@@ -5,19 +5,28 @@ package pParse
 /** Token for base 32 or duatricesimal natural number. All DecimalTokens are also legal base32 tokens, as well as all raw hexadecimal tokens. Base 32
  * , '0' .. '9', followed by 'a' .. 'v' with letter 'o' replaced with 'w'. Some Alpha numeric terms are valid TrigDualInts. */
 trait NatBase32Token extends NatToken
-{ override def subTypeStr: String = "TrigDual"
-
-  /** USes the digitStr to calculate the natural number from the digitStr. */
+{
+  /** Uses the digitStr to calculate the natural number from the digitStr. */
   def asbase32: Int = ???
 }
 
-/** A raw natural integer token for base32 that is not a valid hexadecimal or decimal number */
-case class NatBase32RawToken(startPosn: TextPosn, srcStr: String) extends NatBase32Token
+/** A raw base 32 natural number token includes valid standard decimal natural number and raw hexadecimal tokens. */
+trait NatRawToken extends NatBase32Token
+
+object NatBase32Token
 {
-  override def digitsStr: String = srcStr
+  /** A raw natural integer token for base32 that is not a valid hexadecimal or decimal number */
+  def apply(startPosnIn: TextPosn, srcStrIn: String): NatRawToken = new NatRawToken
+  {
+    val startPosn: TextPosn = startPosnIn
+    val srcStr: String = srcStrIn
+    override def subTypeStr: String = "Base32Raw"
+    override def digitsStr: String = srcStr
+  }
 }
 
+/** An unambiguous base32 natural number token, starts with the 0z characters. */
 case class Nat0tToken(startPosn: TextPosn, digitsStr: String) extends NatBase32Token
-{
+{ override def subTypeStr: String = "NatOt"
   override def srcStr: String = "0t" + digitsStr
 }
